+++ sdk/tables/azure-data-tables/test/ut/serializers_test.hpp
@@ -3,7 +3,7 @@
 // Licensed under the MIT License.
 
 #include "../src/private/serializers.hpp"
-#include "azure/data/tables/tables_client.hpp"
+#include "azure/data/tables/table_client.hpp"
 #include "test/ut/test_base.hpp"
 
 namespace Azure { namespace Data { namespace Test {
