+++ sdk/tables/azure-data-tables/test/ut/transactions_test.hpp
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-#include "azure/data/tables/tables_client.hpp"
+#include "azure/data/tables/table_client.hpp"
 #include "test/ut/test_base.hpp"
 
 namespace Azure { namespace Data { namespace Test {
